Added Routes
Please double check in the html-routes file that I have the right file name. Also, take note of the API URLs so you can properly reference them in other files. Since we're using different tables for different aspects of the character, we'll have to do a few different API calls depending on the data.